refactor(ui): split Details.tsx into per-entity pages + shared ListSection
Details.tsx (1902 lines, 8 detail views) becomes a 26-line entry point
re-exporting per-entity pages under pages/details/ (Cluster, Namespace,
Workload, Pod, Node, Service, Ingress, PersistentVolumes). Exports are
unchanged, so App.tsx routes and Details.test.tsx work unmodified.

New components/ListSection.tsx factors the boilerplate the nine paginated
sections repeated (SectionTitle + Paginator + loading/error/empty +
entities table) behind a column config; it takes the usePagedList state
so sections can still derive side content from the current page.
Group-by/impact sections and non-paginated tables keep their bespoke
markup. Rendered output is unchanged.

UI suite: 338 tests green (333 baseline + 5 new ListSection tests),
tsc clean, production build OK.
